feat: Introduce common crate for shared code
A new `common` crate has been created to house code that is shared
between different parts of the project.

The initial contributions to this crate are the `make_flags` macro and
the VGA driver. The `bootloader` and `kernel` have been updated to use
this shared crate. This change will allow for better code reuse and
organization as the project grows.