feat: Add serial communication driver
This commit introduces a basic serial communication driver for the COM1
port.

The driver is located in the new `common/src/serial.rs` file and
provides functionality for initializing the serial port and transmitting
data. This will be used for logging and debugging purposes.

The `serial.txt` file, which is used to capture serial output, has been
added to the `.gitignore` file.